+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [2.0.2] - 2024-09-29
+### Added
+- faultable module
 ### Changed
 - Simplified status variable check
 - Simplified context merge